auth模块统一错误处理，swagger、参数校验添加

+++ common/src/main/java/com/tuyoo/framework/grow/common/entities/IErrorCode.java
@@ -5,7 +5,6 @@ package com.tuyoo.framework.grow.common.entities;
  * Created by macro on 2019/4/19.
  */
 public interface IErrorCode {
-    long getCode();
-
+    Integer getCode();
     String getMessage();
 }
